fix: allow loading of memtrie during catchup, for shards pending resharding (#12679)
This PR contains a small code change to fix a failure in the tests
`test_resharding_v3_load_mem_trie_`.

In the test a client that doesn't track the parent shard is assigned to
track one of the children in the next epoch after resharding. However,
in order to track the children shard, the client must state sync the
state of the parent and perform the shard split.
This fix allows the catchup routine to build the memtrie even with
`load_mem_tries_for_tracked_shards=false`, only if the shard is "pending
resharding in the future".

The first two commits are updates of code comments.