feat(state-dumper): rewrite the state dumper logic (#12492)
The current state dumper code is sort of difficult to follow, and
doesn't make good use of the available cores to obtain and upload parts.
It starts one thread per shard that dumps one part on each iteration of
a big loop (that includes a good amount of unnecessary/redundant lookups
and calculations). So here we rewrite the logic so that instead of
starting one thread per shard and looping over part IDs like that, we
just figure out what parts need to be dumped when we see a new epoch,
and then spawn futures to obtain and upload the parts. Now the part
upload speed will be limited by the number of allowed "obtain part"
tasks (4), and the speed of generating those parts.

This has the advantage of not needing to change anything to work with
dynamic resharding, and the part upload is much faster. On a forknet run
with recent mainnet state, the old dumper takes around an hour and a
half to dump all the parts, and this version takes around half an hour
(could maybe be improved by tweaking/making configurable the number of
allowed tasks obtaining parts at a time)

This could be refactored further because there's still some leftover
structures from the previous implementation that don't fit super
cleanly, but this can be done in a future PR.