test(state-sync): add test case for forks of the sync block (#12698)
This adds a test that currently fails because the state sync block
doesn't handle the case where the first sync block seen doesn't end up
finalized. We add a drop_blocks_by_height() function to the builder that
will have the network layer drop any block messages for the given
height. The result of this is that the block producer for the sync block
will start state syncing with that sync_hash, but no other nodes will
ever see that block, so they'll all know about a different sync_hash.
This test case relies on the fact that this block producer will need to
state sync in that epoch. This happens to be true, but in the future it
would be good to more explicitly define what shards will be tracked up
front.