[store] Introduce Chain store adapter (#12159)
For now this PR only moves some of the functionality from ChainStore to
ChainStoreAdapter, the simpler functions.

We also do away with the cache in ChainStore which was originally
disabled in https://github.com/near/nearcore/pull/12459

In future PRs, will work towards getting a better suited interface with
helper functions which are too complex to be a part of
ChainStoreAdapter.

Also big issue is for the ChainStoreUpdateAdapter seeing a consistent
view of all the updates. Will need to see how to make that happen!