fix(state-sync): set the sync hash only when it's final (#12720)
There's currently a bug that can happen if the sync hash that we set
doesn't end up on the canonical chain. Here we add a couple more test
cases for that, and fix it by considering only finalized blocks when we
set the state sync hash, and changing the snapshot logic to check
whether the current head is the prev block of what will be the sync hash
block. This means the sync hash is found a little later than it was
before, but on a chain with long epoch lengths, it shouldn't be a
problem.