[store] Save genesis height in DB for ChainStoreAdapter (#12828)
Part 1 of removing dependency of ChainStoreAdapter from genesis_height

This PR saves the genesis height in MISC column of DB under the key
`GENESIS_HEIGHT_KEY`. We already store the genesis hash and the genesis
state roots.

This is useful later in the chain store adapter where we would no longer
need to pass in the genesis_height as a parameter.

In the current PR we are just comparing the passed in param with the
stored value. In part 2 of the PR, we would remove this param.