fix(state-dumper): continue dumping state for other shards upon error (#12869)
Currently the part upload tasks (shard_id, part_id) are enumerated in a
Vec that we shuffle and take from with buffer_unordered(). But if
there's a shard that can't be dumped because obtain_state_part() is
giving an error, and this happens for enough parts in the shard, then
this stream of tasks will no longer start any new part upload tasks,
even for shards that have no issues.

Fix it by having one stream of tasks per shard so that each one will be
allowed to complete even if there are errors in the others. This also
simplifies things somewhat since there's no need to keep the
`oneshot::Sender` in a `HashMap`