native transactions benchmark bypassing the json-rpc layer (#12899)
Implements the TxGenerator actor, residing in the `neard` at the same
level as `jrpc-rpc`, `ClientActor` and the other top-level entities.
When started tries to read the config and load the account details from
the filesystem. That's one way to get the account private keys required
to sign the transaction. If successful enters the (currently infinite)
loop generating the native token transactions and feeding them directly
into `ClientActor`.
Augments the logs with lines like below:
```
2025-02-14T09:18:56.689941Z  INFO transaction-generator: total="Stats { pool_accepted: 303919, pool_rejected: 1, processed: 297267, failed: 0}"
2025-02-14T09:18:56.689951Z  INFO transaction-generator: diff="Stats { pool_accepted: 6651, pool_rejected: 0, processed: 3157, failed: 0 }"
```
The feature "tx_generator" is introduced to select whether the `neard`
should be built with 'TxGenerator' support.

Tested:
run `just do-it 5000`.